back/requirements.txt
anyio==3.6.2
asyncpg==0.27.0
click==8.1.3
dnspython==2.2.1
email-validator==1.3.0
fastapi==0.88.0
greenlet==2.0.1
h11==0.14.0
idna==3.4
psycopg2-binary==2.9.5
pydantic==1.10.2
PyJWT==2.6.0
python-dotenv==0.21.0
sniffio==1.3.0
SQLAlchemy==1.4.44
starlette==0.22.0
typing-extensions==4.4.0
uvicorn==0.20.0
